requirements.txt
# Deployment and work
pip>=9.0.1
Django==1.8.12
psycopg2
path.py==2.4.1
ipdb==0.10.2
ipython==5.3.0

# Deployment
gunicorn==19.3.0
django-sslify>=0.2

# tests
mock
freezegun
factory_boy
coveralls
flake8

djangorestframework
djoser
newrelic

# Admin skin
django-grappelli

# coffin and django-jinja --- can't find, probably deprecated
# coffin
# Was used for macros
# django-jinja==1.4.1
# jinja2

# Compresses linked and inline JavaScript or CSS into a single cached file.
django-compressor==1.5

# For distant static files
django-assets
boto
django-storages==1.1.8

jsmin

# oauth2
# -e git+https://github.com/jgorset/fandjango.git#egg=fandjango
# fandjango
python-openid

# -e git+https://github.com/martinrusev/django-redis-sessions.git#egg=django-redis-sessions
django-redis-sessions
-e git+https://github.com/tgrf/django-constance.git#egg=django-constance
django-picklefield
celery==3.1.24
django-celery-with-redis==3.0
django-celery==3.1.17
dj-database-url==0.2.1
pylibmc
django-pylibmc-sasl==0.2.4
hiredis

# dates
python-dateutil==1.5

Pillow==3.2.0
requests==2.0.0
jsonfield

unidecode

# connecting with FB, TT
# https://github.com/python-social-auth/social-app-django
social-auth-app-django
django-twitter-api==0.2.10
twitter-api==0.1.5

# for tags
django-taggit
django-taggit-autosuggest

# for image resizing
django-imagekit

# configured in django-admin fields: description and outcome_reason
django-wysiwyg==0.8.0
django-ckeditor==5.1.1

# search
# elasticsearch>=1.0.0,<2.0.0
# django-haystack>=1.0.0,<2.0.0
elasticsearch
django-haystack
git+https://github.com/liberation/django_elasticsearch.git
celery-haystack

# certifi for search
certifi==0.0.8